% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_binary.R
\name{calc_prob_binary_unnorm}
\alias{calc_prob_binary_unnorm}
\title{Calculate unnormalized probability for binary vector}
\usage{
calc_prob_binary_unnorm(binaryvec, betamatrix)
}
\arguments{
\item{binaryvec}{a vector of zeros and ones}

\item{betamatrix}{a matrix of beta coefficients}
}
\description{
Calculate unnormalized probability for binary vector
}

