% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_binary.R
\name{int_to_bin}
\alias{int_to_bin}
\title{Converts integer to vector of zeros and ones}
\usage{
int_to_bin(x)
}
\arguments{
\item{x}{a numeric vector of positive integers}
}
\value{
a matrix in which each column corresponds to one integer of x vector
}
\description{
Converts integer to vector of zeros and ones
}

