% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_binary.R
\name{sample_binary}
\alias{sample_binary}
\title{Draw a binary matrix}
\usage{
sample_binary(beta, G, tmax = 300)
}
\arguments{
\item{beta}{a beta matrix}

\item{G}{a graph matrix}

\item{tmax}{number of vectors to draw}
}
\description{
Draw a binary matrix
}

