% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_trinary.R
\name{sample_trinary}
\alias{sample_trinary}
\title{Sample a trinary matrix}
\usage{
sample_trinary(binary, pp)
}
\arguments{
\item{binary}{a binary matrix}

\item{pp}{a probability vector}
}
\description{
Sample a trinary matrix
}

