% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_alpha.R
\name{update_alpha}
\alias{update_alpha}
\title{Update the alpha vector, the vector of sample-specific random effects}
\usage{
update_alpha(alpha, y, trinary)
}
\arguments{
\item{alpha}{vector of sample-specific random effects}

\item{y}{data matrix}

\item{trinary}{trinary indicator matrix}
}
\description{
Update the alpha vector, the vector of sample-specific random effects
}

