% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_k.R
\name{update_k}
\alias{update_k}
\title{Update kminus or kplus vector}
\usage{
update_k(k, indic = 1, trinary, y, shape = 10, scale = 0.02,
  sd_jump = 0.1)
}
\arguments{
\item{k}{a vector of kminus or kplus values}

\item{indic}{indicator of whether to update kminus (-1) or kplus (1)}

\item{trinary}{a trinary matrix}

\item{y}{data matrix}

\item{shape}{prior shape parameter for inv gamma}

\item{scale}{prior scale parameter for inv gamma}
}
\description{
Update kminus or kplus vector
}

