% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mu.R
\name{update_mu}
\alias{update_mu}
\title{Update the mean vector mu (after Xu et al., 2012)}
\usage{
update_mu(mu, trinary, y, alpha, tau = 1, sd_jump = 0.1)
}
\arguments{
\item{mu}{a mean vector}

\item{trinary}{a trinary matrix}

\item{tau}{prior sd of mu}

\item{sd_jump}{sd of jump distribution}
}
\description{
Update the mean vector mu (after Xu et al., 2012)
}

