% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pp.R
\name{update_pp}
\alias{update_pp}
\title{Update the bernoulli parameter pp vector}
\usage{
update_pp(trinary)
}
\arguments{
\item{trinary}{trinary matrix}
}
\description{
Update the bernoulli parameter pp vector
}

