% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_trinary.R
\name{update_trinary}
\alias{update_trinary}
\title{Update trinary matrix}
\usage{
update_trinary(binary, y, pp, mu, sigma, kplus, alpha)
}
\arguments{
\item{binary}{a binary matrix of zeros and ones}

\item{y}{a data matrix}

\item{pp}{a vector of probabilities}

\item{mu}{a vector of means}

\item{sigma}{a vector of standard deviations}

\item{kplus}{a vector of kplus values}
}
\description{
Update trinary matrix
}

